requirements.txt
praw>=5.2.0
mongoengine
pyyaml
